Refactor: /item page
