+++ src/components/GNB.jsx
@@ -1,4 +1,6 @@
 import { Link } from "react-router-dom";
+import Logo from "@/assets/images/logo.svg";
+import LogoTypoOnly from "@/assets/images/logo_typo_only.svg";
 
 function GNB() {
   return (
@@ -6,8 +8,8 @@ function GNB() {
       <div className="container flex flex-between">
         <Link to="/" className="logo">
           <picture>
-            <source srcSet="/images/logo.svg" media="(min-width:768px)" />
-            <img src="/images/logo_typo_only.svg" alt="판다마켓 로고" />
+            <source srcSet={Logo} media="(min-width:768px)" />
+            <img src={LogoTypoOnly} alt="판다마켓 로고" />
           </picture>
         </Link>
         <Link to="/signup" className="button button--small-48">
